#93 - Refactored Building Icons
1-Deletion of file building_marker.dart
Initially used for returning a marker with the ontap bottomsheet. It was decided that the contents should be split up in order to change the icon image. Currently, we will put the bottomsheet into a new file called bottomsheet_widget.dart, and have the marker creation inside of the map_widget.dart file. The icon image for the marker will use formAssetImage and import an image from the asset folder.

2- Added bottomsheet_widget.dart
The ontap for the building markers will return a bottomSheet_widget. The bottomsheet_widget.dart will have all the bottomSheet information needed when the user taps on the building icon.

3- Edit of map_widget.dart
Added a BitmapDescriptor buildingIcon to hold the icon image. The function setBuildingIcons will put the image into buildingIcon and this function will be called at initialisation.

4- Current Progress
The icon for buildings is currently the hall building name. It is shown on every building. More changes to come.