Fix #107 - Add support for classrooms as a destination
This commit adds two constant destinations that are to be chosen depending on the classroom:
- The H building
- The JMSB building

Note: Users can still navigate to campuses by explicitly selecting them, e.g., Loyola.

The filtering works by checking the first letter of the chosen room.
The destination is set based on that first letter.

A JSON file containing the supported destinations was added under `assets/destinations.json`.