#35 - Added outdoor poi files
1- POI json file added
file has 4 poi from loyola and 5 poi from sgw

2- Added outdoor_poi.dart file
This file is considered the outdoor poi object. For now, the required attributes are marked, but could change.

3- Added outdoor_poi_list.dart file
This file will be used to read the poi json file and create the outdoor poi objects. Unfinished. Missing the parsing of the file and returning a list of the poi objects.