#35 - Added POI cards in outdoor poi screen
1- Updated outdoor_poi models
Added a return list for the outdoor_poi_list. Reading the json file method still unfinished.

2- Added listview of cards in the outdoor_interest
The outdoor screen's structure is created. Can only be viewed once the outdoor poi json file is read and the list of outdoor poi objects are created. Design of the cards and of the page will follow.