fix #93 - Unique Marker Icon on each Building
1- Changes on map_widget.dart
The set buildingList is instantiated at the beginning from creating the BuildingList object. The list of icons cause issues where the icons would not show, therefore for now, the 10 paths are put into the set iconSet. May come back in the future to refactor this part.

2- campus_building_info, buildingInformation
Readded the icon image paths and variable. For now they aren't used due to issue related above. Will revisit if there is time in the future.

3- building_list.dart
Now readingBuildingFile is called in the constructor. All other functions are now private since they are never needed outside of this class.