#93 - Refactored Code for Icons to allow resizing
Added method getBytesFromAsset and edited setBuildingIcons
This allows changing the size of the icon markers on the map.
Previous method was assumed by default size 500, currently at 350.
At 500 it was initially unclear on the edges of the marker due to the
marker icon being white and map background being very light colored.
At 350, the marker icon is contained within the building outlines.