#120 - Add Weekday widget
The Weekday widget contains a Text element along with a list of courses represented by a ListTile.

Each course is represented by a ListTile showing the course summary, time, and location (if any).
A directions button is activated if the location is set.

Currently, the default destination upon pressing the button is the SGW campus.

What's left to do is to check whether this location is in our list of [supported locations](assets/destinations.json)

We may add a Course widget to allow for more granularity in the design.