#35 - Added Image Placeholder
Image Logo for the poi now have the template on file outdoor_interest.
Need to be updated for each poi with their respective logo or image.
Currently all have the mcdonald's logo. Also need to be styled
appropriately.

pointsofinterest.json file was updated with the correct lng values.
Some of them did not have the negative value. Also added open and close
time for park.

Also changed the OutdoorPOIList name to more meaningful name in the
outdoor_interest.dart file.