#35 - Created 2 tabs for POI and reorganized code
1- New POI widget file poi_display.dart
This widget holds the list of the poi

2- Tabs in outdoor_interest.dart
There are now 2 tabs to list the poi for the respective campus.
The old list that displays the poi was moved to the new POI widget file
called poi_display.dart. It now only needs the snapshot and the campus
string. A back button was added on the app bar. Direction buttons were
added.

Missing styling of the widgets.