#34 - Add fromJson() + refactor
This commit changes how Buildings are represented.
Buildings now include their edges (outline) shown on the map.
To show a building on the map, it's sufficient to define GoogleMap's `polygons` attribute to be that of a building.
As it is necessary to show all buildings, a set of building polygons is made containing every Building's polygons.