#34 - Add initial support to only draw on requested floor
Painter service will now only draw on the correct floor.
For example, of the requested path is on the 8th floor in Hall building, meaning direction from H803 to H805,
the path will only be drawn on the 8th floor indoor widget.
Same rules apply to other floors and buildings.