#166 - Add search service
This service's main function `query(String)` searches for an element in the list of supported elements for navigation.

The text parameter should be upper cased before it is used as every element's string is upper cased for comparison.

The `indoor(Reachable)` function searches for an indoor location in the list of supported locations and returns its .toLatLng().