#166 - Use indoor's building as outdoor destination
This commit makes it feasible to select an indoor location all the while seeing an outdoor itinerary on the map.

This is done by using class hierarchy wherein an indoor location's grand parent object (a Building) is used in the object's .toLatLng() function (since indoor location implement the Reachable interface - they must define this function).

Some of the logic may be faulty, but it does the job of showing an itinerary from an indoor selection.